refactor(polytope_playground): extract wireframe edge geometry into module
Moves the stereographic clip/reconstruction, chord and great-circle-arc edge builders, and per-cell w-slice helpers (~760 lines, consts included) out of main.rs into wireframe_geom.rs. Pure move; comments leaned in a later pass.
